+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/service/PublicationServiceTest.java
@@ -13,7 +13,6 @@ import uk.gov.hmcts.reform.pip.data.management.database.AzureBlobService;
 import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
 import uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper;
 import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
-import uk.gov.hmcts.reform.pip.data.management.models.location.LocationType;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Language;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
@@ -22,7 +21,6 @@ import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactTriggerS
 import uk.gov.hmcts.reform.pip.data.management.utils.PayloadExtractor;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -40,7 +38,6 @@ import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.ARTEFACT_ID;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.CONTENT_DATE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.FILE;
-import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_TYPE_MATCH;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_VENUE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.MANUAL_UPLOAD_PROVENANCE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.NO_COURT_EXISTS_IN_REFERENCE_DATA;
@@ -56,7 +53,6 @@ import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTe
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.TEST_KEY;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.TEST_VALUE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.VALIDATION_ARTEFACT_NOT_MATCH;
-import static uk.gov.hmcts.reform.pip.data.management.helpers.ConstantsTestHelper.MESSAGES_MATCH;
 
 @SuppressWarnings({"PMD.ExcessiveImports"})
 @ExtendWith(MockitoExtension.class)
@@ -374,58 +370,6 @@ class PublicationServiceTest {
 
         assertEquals(LocalDateTime.now().plusDays(7).toLocalDate(), returnedArtefact.getExpiryDate().toLocalDate(),
                      "Expiry date not set correctly for SJP press list");
-    }
-
-    @Test
-    void testArtefactCountService() {
-        when(artefactRepository.countArtefactsByLocation()).thenReturn(List.of("1,3","2,4", "3,6"));
-        assertEquals("location,count\n1,3\n2,4\n3,6\n", publicationService.countArtefactsByLocation(),
-                     MESSAGES_MATCH);
-    }
-
-    @Test
-    void testGetLocationTypeVenue() {
-        List<ListType> venueListTypes = new ArrayList<>();
-        venueListTypes.add(ListType.CROWN_DAILY_LIST);
-        venueListTypes.add(ListType.CROWN_FIRM_LIST);
-        venueListTypes.add(ListType.CROWN_WARNED_LIST);
-        venueListTypes.add(ListType.MAGISTRATES_PUBLIC_LIST);
-        venueListTypes.add(ListType.MAGISTRATES_STANDARD_LIST);
-        venueListTypes.add(ListType.CIVIL_DAILY_CAUSE_LIST);
-        venueListTypes.add(ListType.FAMILY_DAILY_CAUSE_LIST);
-        venueListTypes.add(ListType.IAC_DAILY_LIST);
-
-        venueListTypes.forEach(listType ->
-                                   assertEquals(LocationType.VENUE,
-                                                publicationService.getLocationType(listType),
-                                                LOCATION_TYPE_MATCH));
-    }
-
-    @Test
-    void testGetLocationTypeOwningHearingLocation() {
-        List<ListType> venueListTypes = new ArrayList<>();
-        venueListTypes.add(ListType.ET_DAILY_LIST);
-        venueListTypes.add(ListType.ET_FORTNIGHTLY_PRESS_LIST);
-
-        venueListTypes.forEach(listType ->
-                                   assertEquals(LocationType.OWNING_HEARING_LOCATION,
-                                                publicationService.getLocationType(listType),
-                                                LOCATION_TYPE_MATCH));
-    }
-
-    @Test
-    void testGetLocationTypeNational() {
-        List<ListType> nationalListTypes = new ArrayList<>();
-        nationalListTypes.add(ListType.SJP_PRESS_LIST);
-        nationalListTypes.add(ListType.SJP_PUBLIC_LIST);
-        nationalListTypes.add(ListType.CARE_STANDARDS_LIST);
-        nationalListTypes.add(ListType.PRIMARY_HEALTH_LIST);
-
-        nationalListTypes.forEach(listType ->
-                                      assertEquals(LocationType.NATIONAL,
-                                                   publicationService.getLocationType(listType),
-                                                   LOCATION_TYPE_MATCH));
-
     }
 
     @Test
