+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/service/artefact/ArtefactServiceTest.java
@@ -14,11 +14,13 @@ import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
 import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.NotFoundException;
 import uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper;
 import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
+import uk.gov.hmcts.reform.pip.data.management.models.location.LocationType;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Sensitivity;
 import uk.gov.hmcts.reform.pip.data.management.service.AccountManagementService;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,6 +30,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.ARTEFACT_ID;
+import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_TYPE_MATCH;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_VENUE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.PAYLOAD;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.PROVENANCE;
@@ -41,6 +44,7 @@ import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTe
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.VALIDATION_NOT_THROWN_MESSAGE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ConstantsTestHelper.MESSAGES_MATCH;
 
+@SuppressWarnings({"PMD.ExcessiveImports"})
 @ExtendWith(MockitoExtension.class)
 class ArtefactServiceTest {
 
@@ -324,6 +328,59 @@ class ArtefactServiceTest {
             .thenReturn(Optional.of(artefactWithIdAndPayloadUrl));
         assertEquals(artefactWithIdAndPayloadUrl, artefactService.getMetadataByArtefactId(ARTEFACT_ID, USER_ID),
                      VALIDATION_ARTEFACT_NOT_MATCH);
+    }
+
+    @Test
+    void testGetLocationTypeVenue() {
+        List<ListType> venueListTypes = new ArrayList<>();
+        venueListTypes.add(ListType.CROWN_DAILY_LIST);
+        venueListTypes.add(ListType.CROWN_FIRM_LIST);
+        venueListTypes.add(ListType.CROWN_WARNED_LIST);
+        venueListTypes.add(ListType.MAGISTRATES_PUBLIC_LIST);
+        venueListTypes.add(ListType.MAGISTRATES_STANDARD_LIST);
+        venueListTypes.add(ListType.CIVIL_DAILY_CAUSE_LIST);
+        venueListTypes.add(ListType.FAMILY_DAILY_CAUSE_LIST);
+        venueListTypes.add(ListType.IAC_DAILY_LIST);
+
+        venueListTypes.forEach(listType ->
+                                   assertEquals(
+                                       LocationType.VENUE,
+                                       artefactService.getLocationType(listType),
+                                       LOCATION_TYPE_MATCH));
+    }
+
+    @Test
+    void testGetLocationTypeOwningHearingLocation() {
+        List<ListType> venueListTypes = new ArrayList<>();
+        venueListTypes.add(ListType.ET_DAILY_LIST);
+        venueListTypes.add(ListType.ET_FORTNIGHTLY_PRESS_LIST);
+
+        venueListTypes.forEach(listType ->
+                                   assertEquals(LocationType.OWNING_HEARING_LOCATION,
+                                                artefactService.getLocationType(listType),
+                                                LOCATION_TYPE_MATCH));
+    }
+
+    @Test
+    void testGetLocationTypeNational() {
+        List<ListType> nationalListTypes = new ArrayList<>();
+        nationalListTypes.add(ListType.SJP_PRESS_LIST);
+        nationalListTypes.add(ListType.SJP_PUBLIC_LIST);
+        nationalListTypes.add(ListType.CARE_STANDARDS_LIST);
+        nationalListTypes.add(ListType.PRIMARY_HEALTH_LIST);
+
+        nationalListTypes.forEach(listType ->
+                                      assertEquals(LocationType.NATIONAL,
+                                                   artefactService.getLocationType(listType),
+                                                   LOCATION_TYPE_MATCH));
+
+    }
+
+    @Test
+    void testArtefactCountService() {
+        when(artefactRepository.countArtefactsByLocation()).thenReturn(List.of("1,3","2,4", "3,6"));
+        assertEquals("location,count\n1,3\n2,4\n3,6\n", artefactService.countArtefactsByLocation(),
+                     MESSAGES_MATCH);
     }
 
 }
